refactor : Service Layer에 DIP 적용
