+++ src/main/java/jungle/HandTris/application/impl/MemberImplement.java
@@ -1,0 +1,15 @@
+package jungle.HandTris.application.impl;
+
+import jakarta.transaction.Transactional;
+import jungle.HandTris.application.service.MemberService;
+import jungle.HandTris.domain.repo.MemberRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@Transactional
+@RequiredArgsConstructor
+public class MemberImplement implements MemberService {
+    private final MemberRepository memberRepository;
+
+}
