+++ src/main/java/jungle/HandTris/application/service/GameService.java
@@ -1,60 +1,19 @@
 package jungle.HandTris.application.service;
 
-import jakarta.transaction.Transactional;
 import jungle.HandTris.domain.Game;
-import jungle.HandTris.domain.exception.GameNotFoundException;
-import jungle.HandTris.domain.repo.GameRepository;
 import jungle.HandTris.presentation.dto.request.GameDetailReq;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.UUID;
 
-@Service
-@Transactional
-@RequiredArgsConstructor
-public class GameService {
-    private final GameRepository gameRepository;
+public interface GameService {
+    List<Game> getGameList();
 
-    public List<Game> getGameList() {
-        return gameRepository.findAllByGameStatusNotPlaying();
-    }
+    UUID createGame(GameDetailReq gameDetailReq);
 
-    public UUID createGame(GameDetailReq gameDetailReq) {
-        /*
-        입력 데이터
-            - 게임 카테고리
-            - 총 인원
-        default값
-            - 참가인원 : 1
-            - uuid : 자동 생성
-            - gameStatus : NON_PLAYING
-         */
-        Game createdGame = new Game(gameDetailReq);
-        return createdGame.getUuid();
-    }
+    Game enterGame(long gameId);
 
-    public Game enterGame(long gameId) {
-        Game game = gameRepository.findById(gameId).orElseThrow(GameNotFoundException::new);
-        game.enter();
-        gameRepository.save(game);
-        return game;
-    }
+    Game exitGame(long gameId);
 
-    public Game exitGame(long gameId) {
-        Game game = gameRepository.findById(gameId).orElseThrow(GameNotFoundException::new);
-        game.exit();
-        if (game.getParticipantCount() == 0) {
-            deleteGame(gameId);
-            return game;
-        }
-        gameRepository.save(game);
-        return game;
-    }
-
-    public void deleteGame(long gameId) {
-        Game game = gameRepository.findById(gameId).orElseThrow(GameNotFoundException::new);
-        gameRepository.deleteById(gameId);
-    }
+    void deleteGame(long gameId);
 }
