+++ src/main/java/jungle/HandTris/application/service/MemberService.java
@@ -1,14 +1,4 @@
 package jungle.HandTris.application.service;
 
-import jakarta.transaction.Transactional;
-import jungle.HandTris.domain.repo.MemberRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@Transactional
-@RequiredArgsConstructor
-public class MemberService {
-    private final MemberRepository memberRepository;
-
+public interface MemberService {
 }
