+++ src/main/java/jungle/HandTris/presentation/TetrisController.java
@@ -1,7 +1,7 @@
-package jungle.HandTris.presentation.controller;
+package jungle.HandTris.presentation;
 
-import jungle.HandTris.presentation.dto.request.TetrisMessageRequest;
 import jungle.HandTris.application.service.TetrisService;
+import jungle.HandTris.presentation.dto.request.TetrisMessageRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.handler.annotation.SendTo;
